+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 ### 🛠  Improvements
 - [322](https://github.com/vegaprotocol/data-node/pull/322) - Update the definition of done and issue templates
 - [351](https://github.com/vegaprotocol/data-node/pull/351) - Update to latest Vega and downgrade to Tendermint `v.34.15`
+- [352](https://github.com/vegaprotocol/data-node/pull/352) - Update to latest Vega
+- [356](https://github.com/vegaprotocol/data-node/pull/356) - Added support for fractional positions
 
 ### 🐛 Fixes
 
@@ -37,6 +39,7 @@
 - [340](https://github.com/vegaprotocol/data-node/pull/340) - Add brokers for the new data stores to support sequential and concurrent event processing 
 - [327](https://github.com/vegaprotocol/data-node/pull/327) - Add balances `sql` store and upgrade `gqlgen`
 - [329](https://github.com/vegaprotocol/data-node/pull/327) - Add orders `sql` store
+- [354](https://github.com/vegaprotocol/data-node/pull/354) - Add network limits store and API
 - [355](https://github.com/vegaprotocol/data-node/pull/355) - Persist trade data to postgres backed store 
 
 ### 🐛 Fixes